ref(src): Move more detailed DebuggerContext type into typings folder
